feat(github/star): Add support for backing up starred GitHub repos
This commit adds support for backing up starred repositories, using a
backup rule with `kind: github/star`.

To achieve this, the src/sources/github_repo.rs file was modified to
support both the https://api.github.com/users/{USER}/repos and the
https://api.github.com/users/{USER}/repos API-endpoint. This is made
possible by introducing separate constructors for `repo` and `star`
which set a new `kind` enum-member. `kind` is used to return the kind
string in `fn kind()` as well as to construct the correct
API-endpoint in `fn load`.

Documentation and examples for `kind: github/star` were added, as well
as appropriate unit-testing.

To cut down on "magic string" usage, the enum used for `kind` is
specified in src/helpers/github.rs and also includes the release kind,
making it possible to use the enum throughout the code, instead of
having to rely on the raw strings.

Additionally, this commit adds on to the documentation in README.md by
providing better documentation of the available tokens in the `filter`
expressions.